Prepare the connector HTTP APIs for versioning - Connector Update (#194547)
Towards:  https://github.com/elastic/response-ops-team/issues/125

Prepares connector update API for versioning.

## To verify:
Create some rules and try to update them via UI

(cherry picked from commit 0adba9f7214a77db23c4aafdfd2b623f20ca0dc2)
