+++ x-pack/plugins/actions/common/routes/connector/apis/update/types/latest.ts
@@ -5,4 +5,4 @@
  * 2.0.
  */
 
-export { getAll, getAllUnsecured, getAllSystemConnectors } from './get_all';
+export * from './v1';
