+++ x-pack/plugins/actions/common/validate_empty_strings.ts
