+++ x-pack/plugins/actions/server/application/connector/methods/update/types/index.ts
@@ -5,4 +5,4 @@
  * 2.0.
  */
 
-export { getAll, getAllUnsecured, getAllSystemConnectors } from './get_all';
+export type { ConnectorUpdateParams } from './types';
