+++ x-pack/plugins/actions/server/lib/index.ts
@@ -35,7 +35,6 @@ export {
   isNotificationExecutionSource,
   ActionExecutionSourceType,
 } from './action_execution_source';
-export { validateEmptyStrings } from './validate_empty_strings';
 export { parseDate } from './parse_date';
 export type { RelatedSavedObjects } from './related_saved_objects';
 export { getBasicAuthHeader, combineHeadersWithBasicAuthHeader } from './get_basic_auth_header';
