+++ x-pack/plugins/actions/server/routes/connector/update/index.ts
@@ -5,4 +5,4 @@
  * 2.0.
  */
 
-export { getAll, getAllUnsecured, getAllSystemConnectors } from './get_all';
+export { updateConnectorRoute } from './update';
