+++ x-pack/plugins/actions/server/routes/connector/update/transforms/transform_update_connector_response/latest.ts
@@ -5,4 +5,4 @@
  * 2.0.
  */
 
-export { getAll, getAllUnsecured, getAllSystemConnectors } from './get_all';
+export { transformUpdateConnectorResponse } from './v1';
