+++ x-pack/plugins/actions/server/routes/connector/update/update.test.ts
@@ -5,14 +5,15 @@
  * 2.0.
  */
 
-import { bodySchema, updateActionRoute } from './update';
+import { updateConnectorRoute } from './update';
 import { httpServiceMock } from '@kbn/core/server/mocks';
-import { licenseStateMock } from '../lib/license_state.mock';
-import { mockHandlerArguments } from './legacy/_mock_handler_arguments';
-import { actionsClientMock } from '../actions_client/actions_client.mock';
-import { verifyAccessAndContext } from './verify_access_and_context';
+import { licenseStateMock } from '../../../lib/license_state.mock';
+import { mockHandlerArguments } from '../../legacy/_mock_handler_arguments';
+import { actionsClientMock } from '../../../actions_client/actions_client.mock';
+import { verifyAccessAndContext } from '../../verify_access_and_context';
+import { updateConnectorBodySchema } from '../../../../common/routes/connector/apis/update';
 
-jest.mock('./verify_access_and_context', () => ({
+jest.mock('../../verify_access_and_context', () => ({
   verifyAccessAndContext: jest.fn(),
 }));
 
@@ -21,12 +22,12 @@ beforeEach(() => {
   (verifyAccessAndContext as jest.Mock).mockImplementation((license, handler) => handler);
 });
 
-describe('updateActionRoute', () => {
+describe('updateConnectorRoute', () => {
   it('updates an action with proper parameters', async () => {
     const licenseState = licenseStateMock.create();
     const router = httpServiceMock.createRouter();
 
-    updateActionRoute(router, licenseState);
+    updateConnectorRoute(router, licenseState);
 
     const [config, handler] = router.put.mock.calls[0];
 
@@ -97,7 +98,7 @@ describe('updateActionRoute', () => {
     const licenseState = licenseStateMock.create();
     const router = httpServiceMock.createRouter();
 
-    updateActionRoute(router, licenseState);
+    updateConnectorRoute(router, licenseState);
 
     const [, handler] = router.put.mock.calls[0];
 
@@ -142,7 +143,7 @@ describe('updateActionRoute', () => {
       throw new Error('OMG');
     });
 
-    updateActionRoute(router, licenseState);
+    updateConnectorRoute(router, licenseState);
 
     const [, handler] = router.put.mock.calls[0];
 
@@ -185,7 +186,7 @@ describe('updateActionRoute', () => {
       config: { foo: true },
       secrets: { key: 'i8oh34yf9783y39' },
     };
-    expect(() => bodySchema.validate(body)).toThrowErrorMatchingInlineSnapshot(
+    expect(() => updateConnectorBodySchema.validate(body)).toThrowErrorMatchingInlineSnapshot(
       `"[name]: value '' is not valid"`
     );
   });
