+++ x-pack/plugins/actions/server/routes/create.ts
@@ -8,11 +8,12 @@
 import { schema } from '@kbn/config-schema';
 import { IRouter } from '@kbn/core/server';
 import { ActionResult, ActionsRequestHandlerContext } from '../types';
-import { ILicenseState, validateEmptyStrings } from '../lib';
+import { ILicenseState } from '../lib';
 import { BASE_ACTION_API_PATH, RewriteRequestCase, RewriteResponseCase } from '../../common';
 import { verifyAccessAndContext } from './verify_access_and_context';
 import { CreateOptions } from '../actions_client';
 import { connectorResponseSchemaV1 } from '../../common/routes/connector/response';
+import { validateEmptyStrings } from '../../common/validate_empty_strings';
 
 export const bodySchema = schema.object({
   name: schema.string({
